+++ src/Sftp/Sftp/Models/SFTPSession.cs
@@ -17,6 +17,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using Microsoft.Azure.Commands.Common.Exceptions;
+using Microsoft.Azure.PowerShell.Cmdlets.Sftp.Common;
 
 namespace Microsoft.Azure.PowerShell.Cmdlets.Sftp.Models
 {
@@ -43,7 +44,7 @@ namespace Microsoft.Azure.PowerShell.Cmdlets.Sftp.Models
             string sshProxyFolder = null,
             string credentialsFolder = null,
             bool yesWithoutPrompt = false,
-            int bufferSizeBytes = 256 * 1024)
+            int bufferSizeBytes = SftpConstants.DefaultBufferSizeBytes)
         {
             _connection = new ConnectionInfo(storageAccount, username, host, port);
             _authFiles = new AuthenticationFiles(publicKeyFile, privateKeyFile, certFile);
@@ -143,7 +144,7 @@ namespace Microsoft.Azure.PowerShell.Cmdlets.Sftp.Models
         /// <summary>
         /// Buffer size in bytes for SFTP file transfers.
         /// </summary>
-        public int BufferSizeBytes { get; set; } = 256 * 1024;
+        public int BufferSizeBytes { get; set; } = SftpConstants.DefaultBufferSizeBytes;
 
         /// <summary>
         /// Resolve connection information like hostname and username.
@@ -223,7 +224,7 @@ namespace Microsoft.Azure.PowerShell.Cmdlets.Sftp.Models
             }
 
             // Buffer size option (sftp -B flag)
-            if (BufferSizeBytes != 256 * 1024)
+            if (BufferSizeBytes != SftpConstants.DefaultBufferSizeBytes)
             {
                 args.AddRange(new[] { "-B", BufferSizeBytes.ToString() });
             }